Updated year selection and metadata tags and detecting front or back
