+++ src/app/(main)/seat-covers/[coverType]/[make]/[model]/[year]/page.tsx
@@ -41,13 +41,12 @@ export type TCarCoverSlugParams = {
 // }
 
 export async function generateMetadata({ params }: { params: TPathParams }) {
-  const productType = deslugify(params.productType);
   const make = deslugify(params.make || '');
   const model = deslugify(params.model || '');
   const year = deslugify(params.year || '');
   return {
-    title: `${year} ${make} ${model} ${productType}, Custom Fit - Coverland`,
-    description: `${year} ${make} ${model} ${productType} ᐉ Coverland ⭐ Free, Same-Day Shipping ✔️ Free Returns & Purchase Protection ✔️ Made from premium quality, heavy-duty materials with a soft inner fabric.`,
+    title: `${year} ${make} ${model} Seat Covers, Custom Fit - Coverland`,
+    description: `${year} ${make} ${model} Seat Covers ᐉ Coverland ⭐ Free, Same-Day Shipping ✔️ Free Returns & Purchase Protection ✔️ Made from premium quality, heavy-duty materials with a soft inner fabric.`,
   };
 }
