+++ src/app/(main)/seat-covers/[coverType]/page.tsx
@@ -12,6 +12,13 @@ export async function generateStaticParams() {
   return [{ coverType: 'leather' }];
 }
 
+export async function generateMetadata({ params }: { params: TPathParams }) {
+  return {
+    title: `Seat Covers, Custom Fit - Coverland`,
+    description: `Seat Covers ᐉ Coverland ⭐ Free, Same-Day Shipping ✔️ Free Returns & Purchase Protection ✔️ Made from premium quality, heavy-duty materials with a soft inner fabric.`,
+  };
+}
+
 const coverTypes = ['leather'];
 export default async function Leatherette({ params }: { params: TPathParams }) {
   if (!coverTypes.includes(params.coverType as string)) {
