+++ src/components/header/Header.tsx
@@ -8,6 +8,7 @@ const coverTypes = [
   { title: 'Car Covers', link: '/car-covers' },
   { title: 'SUV Covers', link: '/suv-covers' },
   { title: 'Truck Covers', link: '/truck-covers' },
+  { title: 'Seat Covers', link: '/seat-covers' },
 ];
 
 function Header() {
